requirements.txt
# Core dependencies
numpy>=1.20.0
opencv-python>=4.6.0.66
pytesseract>=0.3.9
Pillow>=9.0.0
pyautogui>=0.9.53
asyncio>=3.4.3
aiohttp>=3.8.0
requests>=2.27.1
beautifulsoup4>=4.11.0

# Optional dependencies (will be used if available)
duckduckgo-search>=3.0.0  # For web search
easyocr>=1.6.0            # For enhanced OCR
mediapipe>=0.8.9          # For gesture/pose detection

# AI/ML Frameworks
tensorflow>=2.5.0         # For object detection/ML capabilities
torch>=1.9.0              # PyTorch for ML models
torchvision>=0.10.0       # For computer vision models
transformers>=4.11.3      # For advanced NLP capabilities
sentence-transformers>=2.2.2  # For text embeddings

# Vector Database & RAG Components
chromadb>=0.4.6          # Local vector database
langchain>=0.0.267       # Framework for RAG and AI chains
faiss-cpu>=1.7.3         # Fast vector similarity search (CPU version)
nest_asyncio>=1.5.5

# Enhanced UI Automation
python-mss>=6.1.0        # Fast screen capture
pynput>=1.7.3            # For keyboard/mouse monitoring

# Data Management
networkx>=2.6.2          # For graph-based relationship modeling
scikit-learn>=0.24.2     # For classical ML algorithms

# Computer Vision
ultralytics>=8.0.0       # For YOLO object detection

# Google Cloud Vision
google-cloud-vision>=3.1.0  # Cloud Vision API client library
google-auth>=2.3.3       # Google authentication
google-api-python-client>=2.0.0
google-cloud-aiplatform>=1.25.0
google-cloud-language>=2.0.0
google-generativeai>=0.3.0

# Generative AI
huggingface-hub>=0.12.0
